requirements.txt
streamlit==1.27.0
elasticsearch==8.9.0
python-dotenv